refact/front/PopupContainer: moves to src/comopnents/popups/PopupContainer.jsx
